refactor: SM-50 피드백을 반영하여 코드 리팩토링
